just need to figure out assigning the foreign keys..

+++ Database/ModelFirstSmartpoolDatabase/ModelFirstSmartPool/PoolDimensions.cs
@@ -15,9 +15,9 @@ namespace ModelFirstSmartPool
     public partial class PoolDimensions
     {
         public int PoolDimensionsId { get; set; }
-        public double Length { get; set; }
-        public double Width { get; set; }
-        public double Depth { get; set; }
+        public int Length { get; set; }
+        public int Width { get; set; }
+        public int Depth { get; set; }
     
         public virtual Pool Pool { get; set; }
     }
